Feature #7291 - GalleryInstanciator refactoring: - implementing ComponentInstancePreDestruction in order to move code from GalleryInstanciator to GalleryInstancePreDestruction - re-activating some unit tests that were into test-waiting folder - deleting GalleryInstanciator
