+++ gallery/gallery-ejb/src/integration-test/java/com/silverpeas/gallery/GalleryWarBuilder.java
@@ -37,6 +37,7 @@ public class GalleryWarBuilder extends BasicCoreWarBuilder {
    */
   protected <T> GalleryWarBuilder(final Class<T> test) {
     super(test);
+    addPackages(true, "com.silverpeas.gallery");
     addMavenDependencies("org.silverpeas.core.ejb-core:formtemplate");
     addMavenDependencies("org.silverpeas.core.ejb-core:searchengine");
     addMavenDependencies("org.silverpeas.core.ejb-core:comment");
