+++ gallery/gallery-ejb/src/integration-test/java/com/silverpeas/gallery/constant/MediaResolutionTest.java
@@ -24,14 +24,12 @@
 
 package com.silverpeas.gallery.constant;
 
-import com.silverpeas.gallery.GalleryComponentSettings;
 import com.silverpeas.gallery.GalleryWarBuilder;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.silverpeas.test.BasicWarBuilder;
 
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.nullValue;
@@ -42,11 +40,7 @@ public class MediaResolutionTest {
 
   @Deployment
   public static Archive<?> createTestArchive() {
-    return GalleryWarBuilder.onWarForTestClass(MediaResolutionTest.class)
-        .testFocusedOn(warBuilder -> {
-          warBuilder.addClasses(GalleryComponentSettings.class);
-          warBuilder.addPackages(true, "com.silverpeas.gallery.constant");
-        }).build();
+    return GalleryWarBuilder.onWarForTestClass(MediaResolutionTest.class).build();
   }
 
   @Test
