+++ gallery/gallery-ejb/src/integration-test/java/com/silverpeas/gallery/model/InternalMediaTest.java
@@ -23,19 +23,17 @@
  */
 package com.silverpeas.gallery.model;
 
-import com.silverpeas.gallery.GalleryComponentSettings;
 import com.silverpeas.gallery.GalleryWarBuilder;
-import com.silverpeas.gallery.MediaUtil;
 import com.silverpeas.gallery.constant.MediaMimeType;
 import com.silverpeas.gallery.constant.MediaType;
+import org.apache.commons.lang3.time.DateUtils;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
-import org.junit.runner.RunWith;
-import org.silverpeas.util.DateUtil;
-import org.apache.commons.lang3.time.DateUtils;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.silverpeas.date.Period;
+import org.silverpeas.util.DateUtil;
 
 import java.util.Date;
 
@@ -47,14 +45,7 @@ public class InternalMediaTest {
 
   @Deployment
   public static Archive<?> createTestArchive() {
-    return GalleryWarBuilder.onWarForTestClass(InternalMediaTest.class).testFocusedOn(warBuilder -> {
-      warBuilder.addClasses(GalleryComponentSettings.class, MediaUtil.class);
-      warBuilder.addPackages(true, "com.silverpeas.gallery.constant");
-      warBuilder.addPackages(true, "com.silverpeas.gallery.media");
-      warBuilder.addPackages(true, "com.silverpeas.gallery.model");
-      warBuilder.addPackages(true, "com.silverpeas.gallery.process");
-      warBuilder.addAsResource("maven.properties");
-    }).build();
+    return GalleryWarBuilder.onWarForTestClass(InternalMediaTest.class).build();
   }
 
   private Date beginDownloadDate = DateUtils.addDays(DateUtil.getDate(), -50);
