+++ gallery/gallery-ejb/src/integration-test/java/com/silverpeas/gallery/model/StreamingTest.java
@@ -23,7 +23,6 @@
  */
 package com.silverpeas.gallery.model;
 
-import com.silverpeas.gallery.GalleryComponentSettings;
 import com.silverpeas.gallery.GalleryWarBuilder;
 import com.silverpeas.gallery.constant.MediaResolution;
 import com.silverpeas.gallery.constant.MediaType;
@@ -47,13 +46,7 @@ public class StreamingTest extends AbstractMediaTest {
 
   @Deployment
   public static Archive<?> createTestArchive() {
-    return GalleryWarBuilder.onWarForTestClass(StreamingTest.class)
-        .testFocusedOn(warBuilder -> {
-          warBuilder.addClasses(GalleryComponentSettings.class);
-          warBuilder.addPackages(true, "com.silverpeas.gallery.constant");
-          warBuilder.addPackages(true, "com.silverpeas.gallery.model");
-          warBuilder.addAsResource("maven.properties");
-        }).build();
+    return GalleryWarBuilder.onWarForTestClass(StreamingTest.class).build();
   }
 
   @Test
