+++ gallery/gallery-ejb/src/integration-test/java/com/silverpeas/gallery/model/VideoTest.java
@@ -23,7 +23,6 @@
  */
 package com.silverpeas.gallery.model;
 
-import com.silverpeas.gallery.GalleryComponentSettings;
 import com.silverpeas.gallery.GalleryWarBuilder;
 import com.silverpeas.gallery.constant.MediaResolution;
 import com.silverpeas.gallery.constant.MediaType;
@@ -48,13 +47,7 @@ public class VideoTest extends AbstractMediaTest {
 
   @Deployment
   public static Archive<?> createTestArchive() {
-    return GalleryWarBuilder.onWarForTestClass(VideoTest.class)
-        .testFocusedOn(warBuilder -> {
-          warBuilder.addClasses(GalleryComponentSettings.class);
-          warBuilder.addPackages(true, "com.silverpeas.gallery.constant");
-          warBuilder.addPackages(true, "com.silverpeas.gallery.model");
-          warBuilder.addAsResource("maven.properties");
-        }).build();
+    return GalleryWarBuilder.onWarForTestClass(VideoTest.class).build();
   }
 
   @Test
