+++ gallery/gallery-ejb/src/test/resources/org/silverpeas/general.properties
@@ -1,0 +1,67 @@
+#
+# Copyright (C) 2000 - 2013 Silverpeas
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# As a special exception to the terms and conditions of version 3.0 of
+# the GPL, you may redistribute this Program in connection with Free/Libre
+# Open Source Software ("FLOSS") applications as described in Silverpeas's
+# FLOSS exception.  You should have recieved a copy of the text describing
+# the FLOSS exception, and it is also available here:
+# "http://www.silverpeas.org/docs/core/legal/floss_exception.html"
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+ApplicationURL=/silverpeas/
+
+#Nom du workspace (configur\u00e9 dans repository.xml)
+webdav.workspace=jackrabbit
+#Mapping de la serlet webdav sans '/'
+webdav.respository=repository
+
+# working directory for components (created with the instanciators)
+uploadsPath=/
+# working directory for the search engine
+uploadsIndexPath=${project.build.directory}/temp/index
+removeLocksOnInit=yes
+# working directory for temporaries files
+tempPath=${project.build.directory}/temp/temp
+
+sessionTimeout = /admin/jsp/SessionTimeout.jsp
+accessForbidden = /admin/jsp/accessForbidden.jsp
+redirectAppInMaintenance = /admin/jsp/appInMaintenance.jsp
+
+RepositoryTypeTemp=Temp
+
+# logo to print in login page and top banner
+logo=/admin/jsp/icons/logo_silverpeasBig.gif
+smallLogo=/admin/jsp/icons/logo_silverpeas.gif
+smallLogoSilverBlue=icons/DomainbarSilverblue/logo.gif
+
+# advanced search with PDC
+# doit \u00eatre \u00e0 true ou false
+advancedSearchWithPDC = true
+
+# Domain visibility parameter
+# 0 (Default) For all visible
+# 1 Domain Silverpeas Users see all but other domain's users just see there own domain
+# 2 All users just see there own domain (except Administrators)
+domainVisibility = 0
+
+# JRE for applet Drag And Drop
+# pathInstallerJre = /weblib/dragAnddrop/jre1.4.2_11.exe
+pathInstallerJre = http://java.sun.com/update/1.5.0/jinstall-1_5_0_11-windows-i586.cab#version=1,5,0
+
+# http server base
+# Needed to override m_sAbsolute value (in front Apache installation case with port forwarding)
+# Example: http://myserver
+httpServerBase =
